requirements.txt
django>=1.7
djangorestframework>=3.0.4
